Implement getter and setter for project attribute data
Implemented getter and setter functions for attribute data, which
should be used to manipulate it. The methods handle serialization and
deserialization between primitives or objects and the JSON field data.
It is also possible to use other places to store attribute data than
the JSON field under the hood, as is currently done for geometry data,
which is stored in an a separate model field.

Previously these things were handled all over the place, which
caused all kinds of problems, for example many different types of
null values being saved.