Handle empty attributes for update_attribute_data
If the attribute data is not yet set in a project then it is None and that will end
up raising an error.

If the data attribute is none or empty, then we can just skip any updates. This will
make it so that even if we send in None as the value it will not empty the attribute_data
which I see more of a fail safe in this case. This needs to be kept in mind however that
you can't do a full reset on the field by just sending in None. And since this feature has
not been talked about, and is a rare use case, I found this solution to be a bit more safe.