Use Slugrelated field choice/multiple choice attributes
Otherwise there will be problems when the section model
serializes a field and tries to use the `identifier` attribute
of a choice, something that would not be available.

If a choice field would be used then only a string representation
of the choice would be returned after drf serialization. Instead
if a related field is used, then the instance is returned after
drf serialization and the section models own serializer can
do it's thing and access the `identifier` attribute of the instance.