Refactor private file storage to support several endpoint views
This changes the behaviour from only using a single endpoint for all private files
to making it possible to use several endpoints by prefixing.

Note that the same prefix is defined multiple times which is not the nicest way
of handling this and something that could be improved in the future by having
a single prefix value defined in an appropriate place.