Validate only current or past attributes but allow all attributes if no validation
If the client asks for validation then we should only validate the current or past phases
as those are the only ones that matter at that point. If the client does not ask for validation
then all phases can be saved.

Note that this means that the client can not save and validate values that are set in the future
but only attributes that are active or past phases.