Use instance checks instead of value type checks for display values
For fieldsets for instance the value type is fieldset but the values inside the fieldsets
can be anything. To not have to fetch the attribute for each of the field set fields we
check the instance type instead. Currently the fields that are checked are fine to check
with instance type instead of value type since there are not multiple different value types
with the same instance type.