KAAV-407 Fix fieldset edit changelogging to only log changed attribute
When editing properties in a fieldset the mechanism used to figure out
old/new values for fieldsets didn't support them at all. Add support for
figuring out old/new values for fieldsets.

Also change the logging of fieldsets to only log the change of that
small part of a fieldset instead of wholesale logging the entire
fieldset.