Add a database backend wrapper with a check before attempting to install postgis (#330)
This wrapper is copied from django\contrib\gis\db\backends\postgis\base.py and is supposed to work exactly as the default postgis backend. The only change is an additional check for if postgis extension is already installed (was causing errors due to lacking permissions in migrate). This same fix is included in Django versions 4.2 and newer. If upgrading Django to 4.2+, this file can be deleted and the default database backend used.